Add `--ignored-keys` argument to tootkeeper (#10)
* Add `--ignored-keys` argument to tootkeeper

This allows a given list of strings to contain "post", which will then be ignored.

One example of such a string is `dmca_address` which contains "post" but with a different meaning.

There might be more such strings in the future, or someone might want to ignore specific strings for whatever reason, so this is more flexible than having a set list.

Signed-off-by: PolyamSpace <117664621+polyamAdmin@users.noreply.github.com>

* Add `ignored-keys` to ReadMe

Signed-off-by: PolyamSpace <117664621+polyamAdmin@users.noreply.github.com>

---------

Signed-off-by: PolyamSpace <117664621+polyamAdmin@users.noreply.github.com>